+++ MemfaultPackages/bort-ota-lib/src/main/java/com/memfault/bort/ota/lib/UpdateActionHandler.kt
@@ -44,4 +44,6 @@ interface UpdateActionHandler {
             scheduleDownload.scheduleDownload(ota)
         }
     }
+
+    fun initialize()
 }