+++ MemfaultPackages/bort-ota-lib/src/test/java/com/memfault/bort/ota/lib/ABUpdateActionHandlerTest.kt
@@ -94,6 +94,7 @@ class ABUpdateActionHandlerTest {
                 otaRulesProvider = otaRulesProvider,
                 settingsProvider = settingsProvider,
             )
+        handler.initialize()
     }
 
     private suspend fun forEachCallback(block: AndroidUpdateEngineCallback.() -> Unit) {